#76 [Feat] Category Entity 삭제, 해당 객체를 enum으로 변환

+++ src/main/java/com/sparta/soomtut/lecture/service/impl/BoardServiceImpl.java
@@ -34,7 +34,7 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     @Transactional(readOnly = true)
-    public Page<LectureResponseDto> getAllPost(Long category, Pageable pageable) {
+    public Page<LectureResponseDto> getAllPost(int category, Pageable pageable) {
 
         if(category == 0 ){
             Page<Lecture> lectures = lectureService.getLectures(pageable);
